codemonk_pro/requirements.txt
Django>=5.0,<6.0
djangorestframework
psycopg2-binary
celery
redis
drf-spectacular
